# backend/openapi.yaml
openapi: 3.1.0
info:
  title: 補助金申請書自動生成API
  version: 1.0.0
  description: 補助金申請書類の自動生成、検証、エクスポートを行うAPI
  contact:
    name: API Support
    email: support@subsidy-app.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-api-version-info:
    deprecated: false
    sunset: null
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.subsidy-app.example.com
    description: Production server
...
paths:
...
  /research/fetch:
    post:
      tags:
        - research
      summary: コンテンツ取得
      operationId: fetchContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  description: 取得対象URL
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /research/summarize:
    post:
      tags:
        - research
      summary: 市場調査の要約＋出典＋データセット
      operationId: summarizeResearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchSummarizeRequest'
      responses:
        '200':
          description: 要約成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchSummarizeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /research/ingest:
    post:
      tags:
        - research
      summary: エビデンス取り込み
      operationId: ingestEvidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content_id
                - application_id
              properties:
                content_id:
                  type: string
                application_id:
                  type: string
      responses:
        '200':
          description: 取り込み成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
...
components:
  schemas:
...
    ResearchSummarizeRequest:
      type: object
      required: [industry, region]
      properties:
        industry:
          type: string
        region:
          type: string
        keywords:
          type: array
          items:
            type: string
    ResearchSummarizeResponse:
      type: object
      properties:
        summary:
          type: string
        sources:
          type: array
          items:
            type: object
            properties:
              title: { type: string }
              publisher: { type: string }
              year: { type: integer }
              url: { type: string, format: uri }
              notes: { type: string }
        datasets:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              unit: { type: string }
              rows:
                type: array
                items:
                  type: array
                  items:
                    oneOf:
                      - type: integer
                      - type: number
...
    FetchResponse:
      type: object
      properties:
        content_id:
          type: string
        url:
          type: string
        content_type:
          type: string
        content:
          type: string
        extracted_at:
          type: string
          format: date-time
    IngestResponse:
      type: object
      properties:
        evidence_id:
          type: string
        content_id:
          type: string
        application_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: リクエスト不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []